Repoman: Merge IsEbuild class into ebuild
This reduces the need to import the pkgs data, it is now an internal class instance variable.
'pkgs' is still needed, but only to assign it the new pkgs dictionary which the scanner
requires for the ebuild loop.
