+++ pym/repoman/modules/scan/ebuild/ebuild.py
@@ -3,8 +3,13 @@
 from __future__ import print_function, unicode_literals
 
 import re
+import stat
+
+from _emerge.Package import Package
+from _emerge.RootConfig import RootConfig
 
 from repoman.modules.scan.scanbase import ScanBase
+from repoman.qa_data import no_exec, allvars
 # import our initialized portage instance
 from repoman._portage import portage
 from portage import os
@@ -19,16 +24,19 @@ class Ebuild(ScanBase):
 		'''Class init
 
 		@param qatracker: QATracker instance
+		@param portdb: portdb instance
 		@param repo_settings: repository settings instance
 		@param vcs_settings: VCSSettings instance
-		@param changed: changes dictionary
 		@param checks: checks dictionary
 		'''
 		super(Ebuild, self).__init__(**kwargs)
 		self.qatracker = kwargs.get('qatracker')
+		self.portdb = kwargs.get('portdb')
 		self.repo_settings = kwargs.get('repo_settings')
 		self.vcs_settings = kwargs.get('vcs_settings')
 		self.checks = kwargs.get('checks')
+		self.root_config = RootConfig(self.repo_settings.repoman_settings,
+			self.repo_settings.trees[self.repo_settings.root], None)
 		self.changed = None
 		self.xpkg = None
 		self.y_ebuild = None
@@ -37,6 +45,7 @@ class Ebuild(ScanBase):
 		self.eapi = None
 		self.inherited = None
 		self.keywords = None
+		self.pkgs = {}
 
 	def _set_paths(self, **kwargs):
 		repolevel = kwargs.get('repolevel')
@@ -93,11 +102,9 @@ class Ebuild(ScanBase):
 	def set_pkg_data(self, **kwargs):
 		'''Sets some classwide data needed for some of the checks
 
-		@param pkgs: the dynamic list of ebuilds
 		@returns: dictionary
 		'''
-		pkgs = kwargs.get('pkgs').get()
-		self.pkg = pkgs[self.y_ebuild]
+		self.pkg = self.pkgs[self.y_ebuild]
 		self.metadata = self.pkg._metadata
 		self.eapi = self.metadata["EAPI"]
 		self.inherited = self.pkg.inherited
@@ -147,6 +154,75 @@ class Ebuild(ScanBase):
 		# update the dynamic data
 		dyn_pkg.set(self.pkg)
 		return False
+
+	def check_isebuild(self, **kwargs):
+		'''Test the file for qualifications that is is an ebuild
+
+		@param checkdirlist: list of files in the current package directory
+		@param checkdir: current package directory path
+		@param xpkg: current package directory being checked
+		@param validity_future: Future instance
+		@returns: dictionary, including {pkgs, can_force}
+		'''
+		checkdirlist = kwargs.get('checkdirlist').get()
+		checkdir = kwargs.get('checkdir')
+		xpkg = kwargs.get('xpkg')
+		fuse = kwargs.get('validity_future')
+		can_force = kwargs.get('can_force')
+		self.continue_ = False
+		ebuildlist = []
+		pkgs = {}
+		for y in checkdirlist:
+			file_is_ebuild = y.endswith(".ebuild")
+			file_should_be_non_executable = y in no_exec or file_is_ebuild
+
+			if file_should_be_non_executable:
+				file_is_executable = stat.S_IMODE(
+					os.stat(os.path.join(checkdir, y)).st_mode) & 0o111
+
+				if file_is_executable:
+					self.qatracker.add_error("file.executable", os.path.join(checkdir, y))
+			if file_is_ebuild:
+				pf = y[:-7]
+				ebuildlist.append(pf)
+				catdir = xpkg.split("/")[0]
+				cpv = "%s/%s" % (catdir, pf)
+				try:
+					myaux = dict(zip(allvars, self.portdb.aux_get(cpv, allvars)))
+				except KeyError:
+					fuse.set(False, ignore_InvalidState=True)
+					self.qatracker.add_error("ebuild.syntax", os.path.join(xpkg, y))
+					continue
+				except IOError:
+					fuse.set(False, ignore_InvalidState=True)
+					self.qatracker.add_error("ebuild.output", os.path.join(xpkg, y))
+					continue
+				if not portage.eapi_is_supported(myaux["EAPI"]):
+					fuse.set(False, ignore_InvalidState=True)
+					self.qatracker.add_error("EAPI.unsupported", os.path.join(xpkg, y))
+					continue
+				pkgs[pf] = Package(
+					cpv=cpv, metadata=myaux, root_config=self.root_config,
+					type_name="ebuild")
+
+		if len(pkgs) != len(ebuildlist):
+			# If we can't access all the metadata then it's totally unsafe to
+			# commit since there's no way to generate a correct Manifest.
+			# Do not try to do any more QA checks on this package since missing
+			# metadata leads to false positives for several checks, and false
+			# positives confuse users.
+			self.continue_ = True
+			can_force.set(False, ignore_InvalidState=True)
+		self.pkgs = pkgs
+		# set our updated data
+		dyn_pkgs = kwargs.get('pkgs')
+		dyn_pkgs.set(pkgs)
+		return self.continue_
+
+	@property
+	def runInPkgs(self):
+		'''Package level scans'''
+		return (True, [self.check_isebuild])
 
 	@property
 	def runInEbuilds(self):
